+++ custom_components/aprilaire_rs485/__init__.py
@@ -31,6 +31,7 @@ _LOGGER = logging.getLogger(__name__)
 
 PLATFORMS: list[Platform] = [
     Platform.BINARY_SENSOR,
+    Platform.BUTTON,
     Platform.CLIMATE,
     Platform.HUMIDIFIER,
     Platform.SENSOR,